+++ gulpfile.js
@@ -7,14 +7,9 @@ var path = require('path');
 var fs = require('fs');
 var format = require('util').format;
 var _ = require('lodash');
-var browserify = require('browserify');
-var watchify = require('watchify');
-var collapser = require('bundle-collapser/plugin');
-var derequire = require('derequire/plugin');
+var webpack = require('webpack-stream');
 var del = require('del');
 var bistre = require('bistre');
-var source = require('vinyl-source-stream');
-var buffer = require('vinyl-buffer');
 
 // Temporary solution until gulp 4
 // https://github.com/gulpjs/gulp/issues/355
@@ -24,6 +19,8 @@ var gulp = require('gulp');
 var $ = require('gulp-load-plugins')();
 
 var pkg = require('./package.json');
+var excluded = require('./tools/excluded');
+var components = require('./components.json');
 
 var config = {
   path: {
@@ -65,12 +62,15 @@ var config = {
     'bb >= 10'
   ],
   uglify: {
+    compress: {
+      warnings: false
+    },
     output: {
       ascii_only: true
     }
   }
 };
-
+var NODE_ENV = process.env.NODE_ENV;
 var dateFormat = 'isoDateTime';
 
 var banner = [
@@ -94,11 +94,27 @@ var preparingData = function() {
   var uiBase = fs.readFileSync('./less/amui.less', fsOptions);
   var widgetsStyle = '';
 
-  var rejectWidgets = ['.DS_Store', 'blank', 'layout2', 'layout3', 'layout4',
-    'container'];
+  var excludedWidgets = ['.DS_Store', 'blank', 'layout2', 'layout3', 'layout4',
+    'container'].concat(excluded.widgets);
   var allWidgets = fs.readdirSync(WIDGET_DIR).filter(function(widget) {
-    return rejectWidgets.indexOf(widget) === -1;
+    return excludedWidgets.indexOf(widget) === -1;
   });
+
+  // 剔除排除配置中不打包的样式
+  var excludedStyleDep = [];
+  var includedStyleDep = [];
+  var getStyleDep = function(type, plugin) {
+    var basename = path.basename(plugin, '.js');
+
+    if (components.js[basename]) {
+      components.js[basename].depStyle.forEach(function(dep) {
+        if (dep.indexOf('ui.') > -1) {
+          type === 'excluded' ?
+            excludedStyleDep.push(dep) : includedStyleDep.push(dep);
+        }
+      });
+    }
+  };
 
   plugins = _.union(config.js.base, fs.readdirSync('./js'));
 
@@ -107,10 +123,16 @@ var preparingData = function() {
   plugins.forEach(function(plugin, i) {
     var basename = path.basename(plugin, '.js');
 
-    if (basename !== 'amazeui' && basename !== 'amazeui.legacy') {
+    if (basename !== 'amazeui' && basename !== 'amazeui.legacy' &&
+      (excluded.plugins.indexOf(basename) === -1)) {
       jsEntry += (basename === 'core' ? 'var UI = ' : '') +
         'require("./' + basename + '");\n';
+
+      getStyleDep('included', plugin);
     }
+  });
+  excluded.plugins.forEach(function(plugin) {
+    getStyleDep('excluded', plugin);
   });
 
   // widgets partial
@@ -128,9 +150,16 @@ var preparingData = function() {
     widgetsStyle += '\r\n// ' + widget + '\r\n';
     widgetsStyle += '@import ".' + srcPath + '.less";' + '\r\n';
     pkg.themes.forEach(function(item, index) {
-      if (!item.hidden && item.name) {
+      if (!item.hidden && item.name && item.name !== 'one') {
         widgetsStyle += '@import ".' + srcPath + '.' + item.name +
           '.less";' + '\r\n';
+      }
+    });
+
+    // 将 widget 依赖的样式推入数组
+    pkg.styleDependencies.forEach(function(file) {
+      if (file.indexOf('ui.') > -1) {
+        includedStyleDep.push(file);
       }
     });
 
@@ -157,23 +186,48 @@ var preparingData = function() {
   // write partials
   fs.writeFileSync(path.join('./vendor/amazeui.hbs.partials.js'), partials);
 
+  // replace excluded style
+  includedStyleDep = _.uniq(includedStyleDep);
+  excludedStyleDep = _.uniq(excludedStyleDep);
+
+  var intersectionDep = _.intersection(includedStyleDep, excludedStyleDep);
+  excludedStyleDep = _.xor(excludedStyleDep, intersectionDep);
+
+  // console.log(includedStyleDep);
+  // console.log(excludedStyleDep);
+
+  excludedStyleDep.forEach(function(dep) {
+    var regExp = new RegExp('(@import "' + dep + '";)');
+    uiBase = uiBase.replace(regExp, '// $1');
+  });
+
   // write less
   fs.writeFileSync('./less/amazeui.less', uiBase + widgetsStyle);
 };
 
 gulp.task('build:preparing', preparingData);
 
-gulp.task('build:clean', function(cb) {
-  del([
+gulp.task('build:clean', function() {
+  return del([
     config.dist.css,
     config.dist.js
-  ], cb);
+  ]);
 });
 
 // Build to dist dir.
 gulp.task('build:less', function() {
   gulp.src(config.path.less)
     .pipe($.header(banner, {pkg: pkg, ver: ''}))
+    .pipe($.plumber({errorHandler: function(err) {
+      // 处理编译less错误提示  防止错误之后gulp任务直接中断
+      // $.notify.onError({
+      //           title:    "编译错误",
+      //           message:  "错误信息: <%= error.message %>",
+      //           sound:    "Bottle"
+      //       })(err);
+      console.log(err);
+      this.emit('end');
+    }}))
     .pipe($.less({
       paths: [
         path.join(__dirname, 'less'),
@@ -185,7 +239,7 @@ gulp.task('build:less', function() {
       }
     }))
     .pipe($.autoprefixer({browsers: config.AUTOPREFIXER_BROWSERS}))
-    .pipe($.replace('//dn-amui.qbox.me/font-awesome/4.3.0/', '../'))
+    .pipe($.replace('//dn-amui.qbox.me/font-awesome/4.5.0/', '../'))
     .pipe(gulp.dest(config.dist.css))
     .pipe($.size({showFiles: true, title: 'source'}))
     // Disable advanced optimizations - selector & property merging, etc.
@@ -204,47 +258,6 @@ gulp.task('build:fonts', function() {
   gulp.src(config.path.fonts)
     .pipe(gulp.dest(config.dist.fonts));
 });
-
-var bundleInit = function() {
-  var b = browserify({
-    entries: './js/amazeui.js',
-    basedir: __dirname,
-    standalone: 'AMUI',
-    paths: ['./js'],
-    cache: {},
-    packageCache: {}
-  });
-
-  if (process.env.NODE_ENV !== 'travisci') {
-    b = watchify(b);
-    b.on('update', function() {
-      bundle(b);
-    });
-  }
-
-  b.plugin(derequire);
-  b.plugin(collapser);
-  b.on('log', $.util.log);
-  bundle(b);
-};
-
-var bundle = function(b) {
-  return b.bundle()
-    .on('error', $.util.log.bind($.util, 'Browserify Error'))
-    .pipe(source('amazeui.js'))
-    .pipe(buffer())
-    .pipe($.replace('{{VERSION}}', pkg.version))
-    .pipe($.header(banner, {pkg: pkg, ver: ''}))
-    .pipe(gulp.dest(config.dist.js))
-    .pipe($.uglify(config.uglify))
-    .pipe($.header(banner, {pkg: pkg, ver: ''}))
-    .pipe($.rename({suffix: '.min'}))
-    .pipe(gulp.dest(config.dist.js))
-    .pipe($.size({showFiles: true, title: 'minified'}))
-    .pipe($.size({showFiles: true, gzip: true, title: 'gzipped'}));
-};
-
-gulp.task('build:js:browserify', bundleInit);
 
 gulp.task('build:js:fuckie', function() {
   return gulp.src('vendor/polyfill/*.js')
@@ -270,9 +283,40 @@ gulp.task('build:js:helper', function() {
     .pipe(gulp.dest(config.dist.js));
 });
 
+gulp.task('build:js:pack', function() {
+  return gulp.src('js/amazeui.js')
+    .pipe(webpack({
+      watch: !(NODE_ENV === 'travisci' || NODE_ENV === 'production'),
+      output: {
+        filename: 'amazeui.js',
+        library: 'AMUI',
+        libraryTarget: 'umd'
+      },
+      externals: [
+        {
+          jquery: {
+            root: 'jQuery',
+            commonjs2: 'jquery',
+            commonjs: 'jquery',
+            amd: 'jquery'
+          }
+        }
+      ]
+    }))
+    .pipe($.replace('{{VERSION}}', pkg.version))
+    .pipe($.header(banner, {pkg: pkg, ver: ''}))
+    .pipe(gulp.dest(config.dist.js))
+    .pipe($.uglify(config.uglify))
+    .pipe($.header(banner, {pkg: pkg, ver: ''}))
+    .pipe($.rename({suffix: '.min'}))
+    .pipe(gulp.dest(config.dist.js))
+    .pipe($.size({showFiles: true, title: 'minified'}))
+    .pipe($.size({showFiles: true, gzip: true, title: 'gzipped'}));
+});
+
 gulp.task('build:js', function(cb) {
   runSequence(
-    ['build:js:browserify', 'build:js:fuckie'],
+    ['build:js:pack', 'build:js:fuckie'],
     ['build:js:helper'],
     cb);
 });